myapp/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.3.6
azure-storage-blob==12.14.1
python-dotenv==1.0.0